pyproject.toml
[tool.poetry]
name = "repl_python3_3muslimsbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-quran-odoa = "^2.0.0"
PyArabic = "^0.6.14"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
